backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
httpx>=0.27.0
pandas>=2.2.0
openpyxl>=3.1.0
python-multipart>=0.0.9
requests>=2.31.0
tqdm>=4.66.0
numpy>=1.26.0
faker>=22.0.0
fhir.resources>=7.0.0
google-cloud-storage>=2.14.0
